// types/Team.ts
export interface Team {
    address?: string;
    city: string;
    country: string;
    gmaps_place_id?: string;
    gmaps_url?: string;
    key: string;
    lat?: number;
    lng?: number;
    location_name?: string;
    motto?: string;
    name: string;
    nickname: string;
    postal_code: string;
    rookie_year: number;
    school_name?: string;
    state_prov: string;
    team_number: number;
    website?: string;
}